pyproject.toml
[tool.poetry]
name = "katewebapp"
version = "0.1.0"
description = ""
authors = ["blowyourbra1n <blowyourbra1n@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.2"
uvicorn = "^0.27.0.post1"
python-dotenv = "^1.0.1"
jinja2 = "^3.1.3"
python-multipart = "^0.0.7"
aiosmtplib = "^3.0.1"
mysql-connector-python = "^8.3.0"
pymysql = "^1.1.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
